requirements.txt
streamlit
mistralai
langdetect
python-docx